+++ CHANGELOG.md
@@ -9,3 +9,8 @@ Create main.py
 Create fLog.py, fConfig.py
 	log to log\main.log,success
 	log to screen,success
+
+# Rev 0.1.2 
+Create fRedis.py
+	using default setting, connect to redis,success
+	set redis key, success